Re-organized utils folders into modules folder and utils folder
