+++ server/utils/UserUtils.js
@@ -1,6 +1,6 @@
 const mongoose = require('../Mongoose.js')
 const UserModel = require('../models/UserModel.js')
-const { hashPassword } = require('../utils/argon.js')
+const { hashPassword } = require('./argon.js')
 const { UserCreateError } = require('../errors/UserErrors.js')
 
 // Returns user by username, otherwise returns null
